+++ ChatApp/App.js
@@ -1,25 +1,11 @@
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, View } from "react-native";
-
-import ChatListItem from "./src/components/ChatListItem";
-
-const chat = {
-  id: "1",
-  user: {
-    image: ("https://i.imgur.com/cFUpyrd.png"),
-    name: "Kamil",
-  },
-  lastMessage: {
-    text: "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
-    createdAt: "04:30",
-  },
-};
+import ChatsScreen from "./src/screens/ChatsScreen";
 
 export default function App() {
   return (
     <View style={styles.container}>
-
-      <ChatListItem chat={chat}/>
+      <ChatsScreen />
 
       <StatusBar style="auto" />
     </View>
